* improved Update.sh to stop if package doesn't exist
git-svn-id: https://svn.exactcode.de/t2/trunk@6511 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc

+++ misc/archive/Update.sh
@@ -11,10 +11,15 @@ fi
 rm -f $$.diff
 pkg=`echo $pkg | tr [A-Z] [a-z]`
 
-./scripts/Create-PkgUpdPatch $pkg-$ver | tee $$.diff
-patch -p1 < $$.diff
-rm -f $$.diff
+if [ -d package/*/$pkg ]; then
+	./scripts/Create-PkgUpdPatch $pkg-$ver | tee $$.diff
+	patch -p1 < $$.diff
+	rm -f $$.diff
 
-./scripts/Download $pkg
-./scripts/Create-CkSumPatch $pkg | patch -p0
+	./scripts/Download $pkg
+	./scripts/Create-CkSumPatch $pkg | patch -p0
+else
+	echo "ERROR: package $pkg doesn't exist"
+	echo package/*/*$pkg*/ | tr ' ' '\n' | grep -v '*'
+fi
